connecting posts and users and authorization checks

+++ prerequisite/controllers/feed.js
@@ -3,6 +3,7 @@ const path = require("path");
 
 const { validationResult } = require("express-validator");
 const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const { page = 1 } = req.query;
@@ -49,32 +50,42 @@ exports.createPosts = (req, res, next) => {
 
   const imageUrl = req.file.path.replace("\\", "/");
   const { title, content } = req.body;
+
+  let creator;
   // console.log(title, content);
 
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: {
-      name: "Akshay",
-    },
+    creator: req.userId, //from the token
   });
 
+  //saving the post in the posts
   post
     .save()
     .then((result) => {
-      console.log(result);
-      // Create post in db
+      //finding the user of the posts
+      return User.findById(req.userId);
+    })
+    //found user
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save(); //saving the posts in the array of posts in the user profile
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
-      next(err); //since async
+      next(err);
     });
 };
 
@@ -125,6 +136,14 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      // //req.userId that we get from the token
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -153,6 +172,14 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+
+      //req.userId that we get from the token
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
+
       // Check logged in user
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
